requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
apache-airflow==2.8.1
apache-airflow-providers-common-io==1.2.0
apache-airflow-providers-common-sql==1.10.1
apache-airflow-providers-ftp==3.7.0
apache-airflow-providers-http==4.9.0
apache-airflow-providers-imap==3.5.0
apache-airflow-providers-sqlite==3.7.0
apispec==6.4.0
argcomplete==3.2.2
asgiref==3.7.2
attrs==23.2.0
Babel==2.14.0
backoff==2.2.1
blinker==1.7.0
cachelib==0.9.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.2
connexion==2.14.2
cron-descriptor==1.4.0
croniter==2.0.1
cryptography==42.0.1
Deprecated==1.2.14
dill==0.3.8
dnspython==2.5.0
docutils==0.20.1
email-validator==1.3.1
Flask==2.2.5
Flask-AppBuilder==4.3.10
Flask-Babel==2.0.0
Flask-Caching==2.1.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
frozenlist==1.4.1
fsspec==2023.12.2
google-re2==1.1
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kafka-python==2.0.2
lazy-object-proxy==1.10.0
limits==3.7.0
linkify-it-py==2.0.2
lockfile==0.12.2
Mako==1.3.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.4
marshmallow==3.20.2
marshmallow-oneofschema==3.1.0
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.4.0
mdurl==0.1.2
multidict==6.0.4
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0        
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
ordered-set==4.1.0
packaging==23.2
pathspec==0.12.1
pendulum==3.0.0
pluggy==1.4.0
prison==0.2.1
protobuf==4.25.2
psutil==5.9.8
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
PyJWT==2.8.0
python-daemon==3.0.1
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==8.0.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.7.0
rich-argparse==1.4.0
rpds-py==0.17.1
setproctitle==1.3.3
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.51
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
text-unidecode==1.3
time-machine==2.13.0
typing_extensions==4.9.0
tzdata==2023.4
uc-micro-py==1.0.2
unicodecsv==0.14.1
universal-pathlib==0.1.4
urllib3==2.1.0
Werkzeug==2.2.3
wrapt==1.16.0
WTForms==3.1.2
yarl==1.9.4
zipp==3.17.0